refactoring: Removed a lot of legacy code
