+++ src/Entity/Allocation.php
@@ -260,106 +260,4 @@ class Allocation extends DomainAllocation
      * @ORM\Column(type="integer", nullable=true)
      */
     protected ?int $secondaryIndicationCode = null;
-
-    /**
-     * TODO: REMOVE after refactoring.
-     *
-     * @ORM\Column(type="integer")
-     */
-    private int $PZC;
-
-    /**
-     * TODO: REMOVE after refactoring.
-     *
-     * @ORM\Column(type="string", length=255)
-     */
-    private string $PZCText;
-
-    /**
-     * TODO: REMOVE after refactoring.
-     *
-     * @ORM\Column(type="integer", nullable=true)
-     */
-    private ?int $SecondaryPZC = null;
-
-    /**
-     * TODO: REMOVE after refactoring.
-     *
-     * @ORM\Column(type="string", length=255)
-     */
-    private string $SecondaryPZCText;
-
-    /**
-     * TODO: REMOVE after refactoring.
-     *
-     * @ORM\Column(type="string", nullable=true)
-     */
-    private ?string $SK = null;
-
-    /**
-     * TODO: REMOVE after refactoring.
-     *
-     * @ORM\Column(type="integer", nullable=true)
-     */
-    private ?int $RMI = null;
-
-    public function getPZC(): int
-    {
-        return $this->PZC;
-    }
-
-    public function setPZC(int $PZC): void
-    {
-        $this->PZC = $PZC;
-    }
-
-    public function getPZCText(): string
-    {
-        return $this->PZCText;
-    }
-
-    public function setPZCText(string $PZCText): void
-    {
-        $this->PZCText = $PZCText;
-    }
-
-    public function getSecondaryPZC(): ?int
-    {
-        return $this->SecondaryPZC;
-    }
-
-    public function setSecondaryPZC(?int $SecondaryPZC): void
-    {
-        $this->SecondaryPZC = $SecondaryPZC;
-    }
-
-    public function getSecondaryPZCText(): string
-    {
-        return $this->SecondaryPZCText;
-    }
-
-    public function setSecondaryPZCText(string $SecondaryPZCText): void
-    {
-        $this->SecondaryPZCText = $SecondaryPZCText;
-    }
-
-    public function getSK(): ?string
-    {
-        return $this->SK;
-    }
-
-    public function setSK(?string $SK): void
-    {
-        $this->SK = $SK;
-    }
-
-    public function getRMI(): ?int
-    {
-        return $this->RMI;
-    }
-
-    public function setRMI(?int $RMI): void
-    {
-        $this->RMI = $RMI;
-    }
 }
